% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_na.R
\name{clean_na}
\alias{clean_na}
\title{Cleaning numeric NAs by replacing them with column medians}
\usage{
clean_na(df)
}
\arguments{
\item{df}{a data frame containing numeric and/or non-numeric colums.}
}
\value{
a data frame with numer NA replaced by median values
}
\description{
Cleaning numeric NAs by replacing them with column medians
}
