% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outlier.R
\name{detect_outlier}
\alias{detect_outlier}
\title{Detect outliers using the IQR rule}
\usage{
detect_outlier(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A logical vector indicating which values are outliers
}
\description{
Detect outliers using the IQR rule
}
