% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_coor_class.R
\name{lp_cor}
\alias{lp_cor}
\title{Creating a custom s3 class for correlation matrices}
\usage{
lp_cor(df)
}
\arguments{
\item{df}{A data frame with numeric columns}
}
\value{
An object of class 'lp_cor'
}
\description{
Computes a correlation matrix for the numeric columns in a dataframe and returns the object in the lp_cor (LiebrechtPackage correlation) class
}
\examples{
lp_cor(iris)
}
