% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_coor_class.R
\name{plot.lp_cor}
\alias{plot.lp_cor}
\title{Plot method for lp_cor objects}
\usage{
\method{plot}{lp_cor}(x, ...)
}
\arguments{
\item{x}{An object of class 'lp_cor'.}

\item{...}{Additional arguments passed to the plot.}
}
\value{
a ggplot2 heatmap of the correaltion matrix.
}
\description{
Plot method for lp_cor objects
}
\examples{
plot(lp_cor(iris))
}
