% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor_heatmap.R
\name{plot_cor_heatmap}
\alias{plot_cor_heatmap}
\title{Plot a correlation heatmap}
\usage{
plot_cor_heatmap(df)
}
\arguments{
\item{df}{A data frame with numeric columns}
}
\value{
a ggplot2 heatmap of the correlation matrix
}
\description{
Plot a correlation heatmap
}
