% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_coor_class.R
\name{print.lp_cor}
\alias{print.lp_cor}
\title{Print method for the lp_cor objects}
\usage{
\method{print}{lp_cor}(x, ...)
}
\arguments{
\item{x}{An object of class 'lp_cor'.}

\item{...}{Not used.}
}
\description{
Print method for the lp_cor objects
}
