% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_num.R
\name{summarize_num}
\alias{summarize_num}
\title{Summarize numeric variables}
\usage{
summarize_num(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A tibble containing summary statistics (min, q1, median, mean, q3, max, and sd) for numeric variables.
}
\description{
Summarize numeric variables
}
\examples{
summarize_num(iris)
}
